+++ apps/routerconsole/java/src/net/i2p/router/web/helpers/GraphHelper.java
@@ -73,7 +73,7 @@ public class GraphHelper extends FormHandler {
         _persistent = _context.getBooleanPropertyDefaultTrue(GraphListener.PROP_PERSISTENT);
         _graphGlow = _context.getBooleanPropertyDefaultTrue(PROP_GLOW);
         _useUtc = _context.getBooleanPropertyDefaultTrue(PROP_UTC);
-        
+
     }
 
     /**
@@ -449,7 +449,7 @@ public class GraphHelper extends FormHandler {
             buf.append("<option value=\"").append(times[i]).append('"');
             if (times[i] == _refreshDelaySeconds) {buf.append(HelperBase.SELECTED);}
             buf.append('>');
-            if (times[i] > 0) {buf.append(DataHelper.formatDuration2(times[i] * 1000));}
+            if (times[i] > 0) {buf.append(DataHelper.formatDuration2((long) times[i] * 1000));}
             else {buf.append(_t("Never"));}
             buf.append("</option>\n");
         }
